src/requirements.txt
O365==2.0.10
pyyaml==5.3.1